feat: added persist- flag

+++ src/main/java/org/tkit/onecx/workspace/domain/services/WorkspaceService.java
@@ -25,6 +25,9 @@ public class WorkspaceService {
         if (workspace == null) {
             return;
         }
+        if (Boolean.TRUE.equals(workspace.getPersist())) {
+            return;
+        }
         imageDAO.deleteQueryByRefId(workspace.getId());
         menuService.deleteAllMenuItemsForWorkspace(workspace.getId());
         workspaceDAO.delete(workspace);
